Kafka: Add support for dynamic partitioning
The Kafka plugin now supports dynamic partition keys, on the same model
as a few other pmacct configuration keys already do, ie in the form of
variables prefixed by `$`. This allows messages to be sent to a
partition which is determined dynamically according to that key, while
only static partitioning (ie to a single partition specified in
configuration) was supported so far.