Kafka plugin: purge Avro schema with dynamic partitioning
When dynamic partitioning is enabled, the Avro schema is now purged to
each partition that is used by the configured topic. This is done
periodically, and configured the same way as with static partitioning.
The only difference is that the schema is sent to each partition instead
of only to the statically configured one.