[WIP] nfacctd: serialize templates (2)
Templates serialization is improved to support pointers to actual
templates fields, that can be stored following either a legacy
template registry or a new template database. This is supported by
directly serializing the data structure that pointer points to, instead
of systematically serializing all of the `tpl` and `ext_db` array fields
of `struct template_cache_entry` and losing references.