[WIP] enable metrics for named plugins
Named plugins are now supported by the internal stats daemon, which now
prefixes a plugin's metrics labels with the configured name when metrics
are initialised, so that statsD can distinguish similar metrics computed
on plugin instances that are different. An important use case for this
can be instantiating a same plugin several times.